refactor: separating logic to improve code mantainance
