+++ src/areEqual.ts
@@ -1,0 +1,3 @@
+export function areEqual(str1: string, str2: string) {
+    return str1 === str2;
+  }
