+++ src/areEqualIgnoreCase.ts
@@ -1,0 +1,3 @@
+export function areEqualIgnoreCase(str1: string, str2: string) {
+    return str1.toLowerCase() === str2.toLowerCase();
+  }
