+++ src/compareManyIgnoreCase.ts
@@ -1,0 +1,10 @@
+export function compareManyIgnoreCase(...args: any[]) {
+    args = [...args];
+    if (args.length === 0) return false;
+    let equal = true;
+    args.forEach(element => {
+      if (args[0].toLowerCase() !== element.toLowerCase()) equal = false;
+    });
+    return equal;
+  }
+
