+++ src/isLowerCase.ts
@@ -1,0 +1,6 @@
+export function isLowerCase(str: string) {
+    if (typeof str !== 'string') {
+      return throwError(str,'string');
+    }
+    return str !== str.toUpperCase() && str === str.toLowerCase();
+  }
