+++ src/reverse.ts
@@ -1,0 +1,10 @@
+export function reverse(str: string, separator:string = '') {
+    if (typeof str !== 'string') {
+      return throwError(str,'string');
+    }
+    return str
+      .trim()
+      .split(separator)
+      .reverse()
+      .join(separator);
+  }
